Add test coverage
